+++ src/compas_rhino/scene/__init__.py
@@ -68,8 +68,8 @@ def clear_rhino(guids=None):
     compas_rhino.clear(guids=guids)
 
 
-@plugin(category="drawing-utils", pluggable_name="redraw", requires=["Rhino"])
-def redraw_rhino():
+@plugin(category="drawing-utils", pluggable_name="after_draw", requires=["Rhino"])
+def after_draw_rhino(drawn_objects):
     compas_rhino.redraw()